Move other products into report
