+++ app/controllers/reports/ndbn_annuals_controller.rb
@@ -24,9 +24,9 @@ class Reports::NdbnAnnualsController < ApplicationController
     @adult_incontinence_report = adult_incontinence_reporter.report
     reporters << adult_incontinence_reporter
 
-
-    # Other Products
-    @other_products = current_organization.items.where(partner_key: other_products_partner_keys).map(&:name)
+    other_products_reporter = Reports::OtherProductsReportService.new(reporter_params)
+    @other_products_report = other_products_reporter.report
+    reporters << other_products_reporter
 
     # Partner Information report values
     @partners = current_organization.partners
@@ -99,23 +99,6 @@ class Reports::NdbnAnnualsController < ApplicationController
     current_organization.partners.map do |partner|
       partner.profile.children.count
     end.sum
-  end
-
-  def base_item_json(key)
-    file = File.read("db/base_items.json")
-    json = JSON.parse(file)
-
-    json[key].map(&:values).map { |keys| keys[0] }
-  end
-
-  def other_products_partner_keys
-    menstrual_supplies = base_item_json("Menstrual Supplies/Items")
-    miscellaneous = base_item_json("Miscellaneous")
-    training_pants = base_item_json("Training Pants")
-    wipes_adult = base_item_json("Wipes - Adults")
-    wipes_children = base_item_json("Wipes - Childrens")
-
-    menstrual_supplies + miscellaneous + training_pants + wipes_adult + wipes_children
   end
 
   def validate_show_params
