+++ app/views/reports/ndbn_annuals/show.html.erb
@@ -144,7 +144,7 @@
 
         <tbody>
           <tr>
-            <td><%= @other_products %></td>
+            <td><%= @other_products_report[:other_products] %></td>
             <td></td>
             <td></td>
             <td></td>
